json-bench/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "json-bench"
version = "0.1.0"
description = "Benchmarking suite for Python JSON serialization libraries"
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "orjson>=3.10",
    "msgspec>=0.18",
    "ujson>=5.10",
    "python-rapidjson>=1.17",
    "msgpack>=1.1",
    "matplotlib>=3.8",
    "pandas>=2.2",
    "tqdm>=4.0",
]

[project.scripts]
json-bench = "json_bench.cli:main"

[dependency-groups]
test = [
    "pytest>=8.3.2",
]

dev = [
    { include-group = "test" },
]

[tool.hatch.build.targets.wheel]
packages = ["src/json_bench"]

[tool.hatch.build.targets.sdist]
packages = ["src/json_bench"]

[tool.pytest.ini_options]
filterwarnings = ["ignore::DeprecationWarning"]
log_level = "INFO"
